@@ -3,81 +3,49 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@codemirror/lang-json": "^0.20.0",
-    "@codemirror/lang-markdown": "^0.20.1",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^16.11.22",
-    "@types/object-hash": "^2.2.1",
-    "@types/react": "^16.8.5",
-    "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^16.8.5",
-    "@types/react-pdf": "^5.0.0",
-    "@uiw/react-codemirror": "~4.7.0",
-    "bootstrap": "^4.6.1",
-    "classnames": "^2.3.1",
+    "@codemirror/lang-json": "^6.0.2",
+    "@codemirror/lang-markdown": "^6.5.0",
+    "@hello-pangea/dnd": "^18.0.1",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.1",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^25.0.3",
+    "@types/object-hash": "^3.0.6",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
+    "@types/react-pdf": "^7.0.0",
+    "@uiw/react-codemirror": "~4.25.4",
+    "bootstrap": "^5.3.8",
+    "classnames": "^2.5.1",
     "fast-json-patch": "^3.1.1",
     "he": "^1.2.0",
-    "js-cookie": "^3.0.1",
-    "katex": "^0.15.1",
+    "js-cookie": "^3.0.5",
+    "katex": "^0.16.27",
     "object-hash": "^3.0.0",
-    "react": "^16.8.5",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "16.8.5",
-    "react-drag-drop-files": "^2.3.7",
-    "react-error-boundary": "^3.1.4",
-    "react-pdf": "^5.0.0",
-    "react-router": "^6.2.1",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "react-select": "^5.3.2",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
+    "react-drag-drop-files": "^3.1.0",
+    "react-error-boundary": "^6.0.1",
+    "react-pdf": "^10.3.0",
+    "react-router": "^7.11.0",
+    "react-router-dom": "^7.11.0",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.10.2",
     "react-split": "^2.0.14",
-    "reactstrap": "^8.10.1",
+    "reactstrap": "^9.2.3",
     "remarkable": "^2.0.1",
-    "swr": "^2.0.3",
-    "ts-jest": "^29.2.5",
-    "typescript": "^5.8.3",
-    "web-vitals": "^2.1.4"
+    "swr": "^2.3.8",
+    "ts-jest": "^29.4.6",
+    "typescript": "^5.9.3",
+    "web-vitals": "^5.1.0"
   },
   "scripts": {
     "start": "craco start --verbose",
-    "build": "craco build",
+    "lint": "eslint -c eslint.config.mjs src/",
+    "build": "yarn run lint && craco build",
     "test": "craco test --watchAll",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "eslint:recommended",
-      "plugin:react/recommended",
-      "plugin:@typescript-eslint/recommended",
-      "plugin:jsx-a11y/strict"
-    ],
-    "plugins": [
-      "react-hooks",
-      "@typescript-eslint",
-      "jsx-a11y"
-    ],
-    "rules": {
-      "no-prototype-builtins": "off",
-      "prefer-const": "warn",
-      "react-hooks/rules-of-hooks": "error",
-      "react-hooks/exhaustive-deps": "warn",
-      "@typescript-eslint/explicit-function-return-type": "off",
-      "@typescript-eslint/no-empty-interface": "warn",
-      "@typescript-eslint/no-inferrable-types": "off",
-      "@typescript-eslint/ban-ts-comment": "warn",
-      "no-case-declarations": "off",
-      "react/jsx-no-target-blank": "off"
-    },
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-      "ecmaFeatures": {
-        "jsx": true
-      }
-    }
   },
   "browserslist": [
     ">0.2%",
