@@ -41,7 +41,7 @@ export function ImageViewer() {
     if (invalid) {
         return <div className={styles.centered}>
             <Alert color="warning">This content does not appear to be an image (PDF, JPEG or GIF).</Alert>
-        </div>
+        </div>;
     }
 
     return <div className={styles.editorWrapper}>
