@@ -56,7 +56,7 @@ export function TextEditor() {
     if (invalid) {
         return <div className={styles.centered}>
             <Alert color="warning">This content does not appear to be text.</Alert>
-        </div>
+        </div>;
     }
 
     return <div className={styles.editorWrapper}>
