@@ -77,8 +77,8 @@ export function buildValuePresenter<V, D extends Content = Content>(
             return <>
                 {component}
                 <div className={`mt-2 ${styles.editButtons}`}>
-                    <Button onClick={(e) => {e.stopPropagation(); cancelChanges()}}>Cancel</Button>
-                    <Button color="primary" onClick={(e) => {e.stopPropagation(); setChanges()}}>Set</Button>
+                    <Button onClick={(e) => {e.stopPropagation(); cancelChanges();}}>Cancel</Button>
+                    <Button color="primary" onClick={(e) => {e.stopPropagation(); setChanges();}}>Set</Button>
                 </div>
             </>;
         }
@@ -114,7 +114,6 @@ export const BaseValue = ({doc, editing, value, set, cancel}: ValueProps<string 
         return <CodeMirror
             ref={codemirror}
             value={value.current}
-            // eslint-disable-next-line jsx-a11y/no-autofocus
             autoFocus={true}
             extensions={[
                 ...encoding,
@@ -141,18 +140,14 @@ export const BaseValuePresenter = buildValuePresenter(
 
 export type ValueRef = React.MutableRefObject<ValuePresenterRef | null>;
 
-export interface ValueWrapperProps {
+export interface ValueWrapperProps extends React.HTMLAttributes<HTMLDivElement> {
     valueRef?: ValueRef;
-    className?: string;
 }
 
-export const ValueWrapper: FunctionComponent<ValueWrapperProps> = ({
-    className,
-    valueRef,
-    children
-}) => {
-    // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-noninteractive-element-interactions
-    return <div role="cell" className={className} onClick={valueRef && ((e) => {
+export const ValueWrapper: FunctionComponent<ValueWrapperProps> = (props) => {
+    const {children, valueRef, ...rest} = props;
+    // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
+    return <div {...rest} role="cell" onClick={valueRef && ((e) => {
         // Only catch clicks that are outside any other element
         // FIXME: This misses clicks in the header, and also takes clicks in odd places like padding and empty spaces of an item.
         if (e.target === e.currentTarget) {
