@@ -50,7 +50,7 @@ export const DropZoneQuestionContext = createContext<{
     isDndQuestion: false,
     isClozeQuestion: false,
     figureMap: {},
-    calculateDZIndexFromFigureId: (id: string) => 0,
+    calculateDZIndexFromFigureId: (_id: string) => 0,
 });
 
 function isParsonsQuestion(doc: Content | null | undefined): doc is IsaacParsonsQuestion {
@@ -99,7 +99,7 @@ export function ItemQuestionPresenter(props: PresenterProps<ItemQuestionType>) {
         if (isDndQuestion(doc)) {
             const f = async () => {
                 // if the number of drop zones has changed, the indexes of figure zones may need to change.
-                const figures = Array.from(Object.entries(figureMap.current))
+                const figures = Array.from(Object.entries(figureMap.current));
                 for (const figure of figures) {
                     const [id, [dropZones, setDropZones]] = figure;
                     const startIndex = extractFigureRegionStartIndex(doc, id);
@@ -108,8 +108,8 @@ export function ItemQuestionPresenter(props: PresenterProps<ItemQuestionType>) {
                     await new Promise(resolve => setTimeout(resolve, 50));
                     console.log("Updated figure", dropZones.map((dz, i) => ({...dz, index: startIndex + i})));
                 }
-                f();
-            }
+            };
+            void f();
         }
     }, [dropZoneCount]);
 
@@ -202,7 +202,7 @@ export function ItemChoicePresenter(props: PresenterProps<ParsonsItem>) {
     const staticItems = isClozeQuestion && allowSubsetMatch && item.id !== NULL_CLOZE_ITEM_ID ? [NULL_CLOZE_ITEM] : [];
 
     const dropdown = <Dropdown toggle={() => setOpen(toggle => !toggle)}
-                               isOpen={isOpen}>
+        isOpen={isOpen}>
         <DropdownToggle outline className={styles.dropdownButton}>
             <ItemRow item={item} />
         </DropdownToggle>
@@ -225,11 +225,11 @@ export function ItemChoicePresenter(props: PresenterProps<ParsonsItem>) {
 
     if (doc.type === "parsonsItem") {
         return <div className={styles.parsonsItem}
-                    style={{borderLeftWidth: `calc(1px + ${(doc.indentation ?? 0) * 1.5}em)`}}>
+            style={{borderLeftWidth: `calc(1px + ${(doc.indentation ?? 0) * 1.5}em)`}}>
             {dropdown}
             <span className={styles.parsonsIndentPresenter}>
                 <MetaItemPresenter {...props} prop="indentation" name="Indent"
-                                   options={indentationOptions} />
+                    options={indentationOptions} />
             </span>
         </div>;
     } else {
@@ -265,7 +265,7 @@ const DropZoneSelector = (props: PresenterProps<DndItem>) => {
             })}
         </DropdownMenu>
     </Dropdown>;
-}
+};
 
 export function DndChoicePresenter(props: PresenterProps<DndItem>) {
     const {doc, update} = props;
