@@ -134,23 +134,23 @@ const choice: RegistryEntry = {
 const accordion: RegistryEntry = {
     name: "Accordion",
     bodyPresenter: AccordionPresenter,
-}
+};
 const horizontal: RegistryEntry = {
     name: "Side-by-side layout",
     className: styles.horizontal,
     bodyPresenter: ListChildrenPresenter,
-}
+};
 const clearfix: RegistryEntry = {
     name: "Clearfix",
     bodyPresenter: () => <div className={"text-center pb-3"}>
         <hr/>
         <i>This block is used to clear floating elements</i>
     </div>,
-}
+};
 const tabs: RegistryEntry = {
     name: "Tabs",
     bodyPresenter: TabsPresenter,
-}
+};
 const hints: RegistryEntry = {
     name: "Hints",
     bodyPresenter: (props) => <TabsPresenter {...props} />,
@@ -222,7 +222,7 @@ const interactiveCodeSnippet: RegistryEntry = {
 const codeTabs: RegistryEntry = {
     name: "Code Tabs",
     bodyPresenter: TabsPresenter,
-}
+};
 const glossaryTerm: RegistryEntry = {
     name: "Glossary term",
     bodyPresenter: GlossaryTermPresenter,
@@ -285,7 +285,7 @@ const isaacQuestionPage: RegistryEntry = {
 const isaacFastTrackQuestionPage: RegistryEntry = {
     ...isaacQuestionPage,
     metadata: [...isaacQuestionPage.metadata ?? [], "level"]
-}
+};
 const isaacConceptPage: RegistryEntry = {
     ...isaacQuestionPage,
     metadata: [...isaacQuestionPage.metadata ?? [], "summary"]
@@ -354,7 +354,7 @@ export const REGISTRY: Record<ContentType, RegistryEntry> = {
     isaacPageFragment: contentPage,
     isaacConceptPage,
     isaacQuestionPage,
-    isaacFastTrackQuestionPage: isaacQuestionPage,
+    isaacFastTrackQuestionPage,
     isaacEventPage,
     isaacBookIndexPage,
     isaacBookDetailPage,
